server/requirements.txt
fastapi>=0.115.8
uvicorn>=0.34.0
psycopg2-binary>=2.9.0
cachetools>=5.5.2
httpx>=0.28.1
starlette>=0.27.0
ollama>=0.4.7
typing-extensions>=4.8.0